% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/readHAC.R
\name{[.HAC}
\alias{[.HAC}
\title{Extract tuples.}
\usage{
\method{[}{HAC}(x, i, ...)
}
\arguments{
\item{x}{HAC object}

\item{i}{Integer vector}

\item{...}{Currently not used}
}
\value{
HAC object
}
\description{
Extract tuples of HAC object.
}
\details{
Extract subset of tuples.
For instance x[1:2] extracts the first two tuples.
Alternatively the method can be indirectly invoked by
the \code{subset} function.
}
\examples{
\dontshow{
hacfile <- system.file("hac","Hac-test_000001.hac",package="readHAC")
x <- readHAC(hacfile)
}
x[1:2]
subset(x, type == 10000)
split(x, x$type)
}

