% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/readHAC.R
\name{readHAC}
\alias{readHAC}
\title{Read HAC data into R.}
\usage{
readHAC(file)
}
\arguments{
\item{file}{File to read.}
}
\value{
HAC object.
}
\description{
Read raw HAC data file
}
\details{
This function reads the binary HAC format and locates the tuples.
}

